reorganize RootLayout files
